only delete binaries if building all packages (not only those in `subset`)

+++ build.R
@@ -48,9 +48,11 @@ writeLines(c(
 
 # delete binaries that have become available on CRAN, or of multiple versions of
 # the same package
-tgz = list.files(dir, '.+_.+[.]tgz$', full.names = TRUE)
-tgz_name = gsub('_.*', '', basename(tgz))
-file.remove(tgz[!(tgz_name %in% pkgs) | duplicated(tgz_name, fromLast = TRUE)])
+if (!file.exists('subset')) {
+  tgz = list.files(dir, '.+_.+[.]tgz$', full.names = TRUE)
+  tgz_name = gsub('_.*', '', basename(tgz))
+  file.remove(tgz[!(tgz_name %in% pkgs) | duplicated(tgz_name, fromLast = TRUE)])
+}
 
 # download source packages that have been updated on CRAN
 if (file.exists(pkg_file <- file.path(dir, 'PACKAGES'))) {
